+++ packages/create-amplify/src/amplify_project_creator.test.ts
@@ -92,10 +92,10 @@ void describe('AmplifyProjectCreator', () => {
     assert.equal(
       logSpy.mock.calls[14].arguments[0],
       grey(
-        `Amplify Gen 2 collects anonymous telemetry data about general usage of the CLI. Participation is optional, and you may opt-out by using ${cyan(
+        `Amplify collects anonymous telemetry data about general usage of the CLI. Participation is optional, and you may opt-out by using ${cyan(
           'npx ampx configure telemetry disable'
         )}. To learn more about telemetry, visit ${underline(
-          blue('https://docs.amplify.aws/gen2/reference/telemetry')
+          blue('https://docs.amplify.aws/react/reference/telemetry')
         )}`
       )
     );
@@ -137,10 +137,10 @@ void describe('AmplifyProjectCreator', () => {
     assert.equal(
       logSpy.mock.calls[14].arguments[0],
       grey(
-        `Amplify Gen 2 collects anonymous telemetry data about general usage of the CLI. Participation is optional, and you may opt-out by using ${cyan(
+        `Amplify collects anonymous telemetry data about general usage of the CLI. Participation is optional, and you may opt-out by using ${cyan(
           'npx ampx configure telemetry disable'
         )}. To learn more about telemetry, visit ${underline(
-          blue('https://docs.amplify.aws/gen2/reference/telemetry')
+          blue('https://docs.amplify.aws/react/reference/telemetry')
         )}`
       )
     );