+++ packages/sandbox/src/file_watching_sandbox.test.ts
@@ -27,7 +27,7 @@ import {
 import { fileURLToPath } from 'url';
 import { BackendIdentifier } from '@aws-amplify/plugin-types';
 import { AmplifyUserError } from '@aws-amplify/platform-core';
-import { SSMClient } from '@aws-sdk/client-ssm';
+import { ParameterNotFound, SSMClient } from '@aws-sdk/client-ssm';
 
 // Watcher mocks
 const unsubscribeMockFn = mock.fn();
@@ -85,9 +85,7 @@ const backendDeployerDestroyMock = mock.method(backendDeployer, 'destroy', () =>
 );
 const region = 'test-region';
 const ssmClientMock = new SSMClient({ region });
-const ssmClientSendMock = mock.fn();
-mock.method(ssmClientMock, 'send', ssmClientSendMock);
-ssmClientSendMock.mock.mockImplementation(() =>
+const ssmClientSendMock = mock.fn(() =>
   Promise.resolve({
     Parameter: {
       Name: CDK_DEFAULT_BOOTSTRAP_VERSION_PARAMETER_NAME,
@@ -95,6 +93,7 @@ ssmClientSendMock.mock.mockImplementation(() =>
     },
   })
 );
+mock.method(ssmClientMock, 'send', ssmClientSendMock);
 const openMock = mock.fn(_open, (url: string) => Promise.resolve(url));
 
 const testPath = path.join('test', 'location');
@@ -155,7 +154,10 @@ void describe('Sandbox to check if region is bootstrapped', () => {
 
   void it('when region has not bootstrapped, then opens console to initiate bootstrap', async () => {
     ssmClientSendMock.mock.mockImplementationOnce(() => {
-      throw new Error('Stack with id CDKToolkit does not exist');
+      throw new ParameterNotFound({
+        $metadata: {},
+        message: 'Parameter not found',
+      });
     });
 
     await sandboxInstance.start({
@@ -465,7 +467,7 @@ void describe('Sandbox using local project name resolver', () => {
     // Mimic BackendDeployer taking 200 ms.
     backendDeployerDeployMock.mock.mockImplementationOnce(async () => {
       await new Promise((resolve) => setTimeout(resolve, 200));
-      return { stdout: '', stderr: '' };
+      return { deploymentTimes: {}, stdout: '', stderr: '' };
     });
 
     // Not awaiting so we can push another file change while deployment is ongoing