+++ packages/sandbox/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @aws-amplify/sandbox
 
+## 1.0.5
+
+### Patch Changes
+
+- 0200d11: Bump baseline CDK version to 2.132.0 to support AWS SDK bundling
+- Updated dependencies [8968cf4]
+- Updated dependencies [4b249ed]
+- Updated dependencies [0200d11]
+- Updated dependencies [20bf679]
+  - @aws-amplify/client-config@1.0.5
+  - @aws-amplify/backend-deployer@1.0.1
+
+## 1.0.4
+
+### Patch Changes
+
+- 4dd9d5a: Fix case when CDK boostrap parameter is not found
+- Updated dependencies [ca9d68d]
+- Updated dependencies [ca9d68d]
+  - @aws-amplify/deployed-backend-client@1.0.2
+  - @aws-amplify/client-config@1.0.4
+
 ## 1.0.3
 
 ### Patch Changes